refactoring : domain 중심으로 변경
