+++ src/app/document/func/text.py
