+++ src/app/document/model.py
@@ -2,9 +2,22 @@
 📄 Document Data Models
 """
 from datetime import datetime
-from typing import Dict, Any, Optional
+from typing import Dict, Any
 from dataclasses import dataclass
 import uuid
+from pydantic import BaseModel
+
+class PDFAnalysisResult(BaseModel):
+    """PDF 분석 결과"""
+    language: str  # 'korean', 'english', 'mixed', 'unknown'
+    has_tables: bool
+    has_images: bool
+    complexity: str  # 'simple', 'medium', 'complex'
+    file_size: int
+    estimated_pages: int
+    text_density: str  # 'low', 'medium', 'high'
+    font_complexity: str  # 'simple', 'complex'
+    recommended_loader: str
 
 
 @dataclass
@@ -42,4 +55,5 @@ class DocumentChunk:
     def __post_init__(self):
         if not self.id:
             self.id = f"{self.document_id}_chunk_{self.chunk_index}"
+
