--- src/app/models/__init__.py
@@ -1,3 +1,0 @@
-"""
-📊 Data Models Package
-"""
