--- src/app/schemas/__init__.py
@@ -1,3 +1,0 @@
-"""
-📋 API Schemas Package
-"""
